// src/types/business-plan.ts
export type DraftPo = {
  id: number;
  createdAt: string;
  lastUpdatedAt: string;
  createdBy: number;
  lastUpdatedBy: number;
  lastProgramUpdate?: string | null;
  code: string;
  status: number;
  lastUpdatedProgram: string;
  contractId: number;
  contractCode?: string | null;
  codeBooking: number;
  customerName: string;
  deliveryLocation: string;
  importCountry: string;
  goodName: string;
  goodId: number;
  deliveryMethod: string;
  paymentMethod: string;
  paymentCurrency: string;
  unitPrice: number;
  unitOfMeasure: string;
  quantity: number;
  goodType: string;
  expectedDelivery: string;
  businessPlanId?: number | null;
};

export type BusinessPlanCostItem = {
  id: number;
  createdAt: string;
  lastUpdatedAt: string;
  createdBy: number;
  lastUpdatedBy: number;
  lastProgramUpdate?: string | null;
  code: string;
  status: number;
  lastUpdatedProgram: string;
  businessPlanId: number;
  totalFreightEbs: number;
  factoryToPortCost: number;
  warehouseRentalCost: number;
  truckingCost: number;
  localCharges: number;
  earlyUnloadingFee: number;
  thcFee: number;
  sealFee: number;
  infrastructureFee: number;
  customsSupervisionFee: number;
  fumigationPerContainer: number;
  fumigationPerLot: number;
  quarantineFee: number;
  ttFee: number;
  coFee: number;
  doFee: number;
  palletFee: number;
  jumboBagFee: number;
  amsFee: number;
  customsTeamFee: number;
  customsReceptionFee: number;
  clearanceCost: number;
  interestCost: number;
  vatInterestCost: number;
  exchangeRateCost: number;
  dhlFee: number;
  brokerageFee: number;
  taxRefundCost: number;
  qcCost: number;
  generalManagementCost: number;
  currency?: string | null;
  notes?: string | null;
  otherCost?: number | null;
};

export type BusinessPlanSupplierItem = {
  id: number;
  createdAt: string;
  lastUpdatedAt: string;
  createdBy: number;
  lastUpdatedBy: number;
  lastProgramUpdate?: string | null;
  code: string;
  status: number;
  lastUpdatedProgram: string;
  businessPlanId: number;
  supplierId: number;
  goodId: number;
  goodName: string;
  region: string;
  quantity: number;
  purchasePrice: number;
  totalAmount: number;
  expectedDeliveryDate?: string | null;
  goodType: string;
  exportPort: string;
};

export type BusinessPlanTransactionInfoItem = {
  id: number;
  createdAt: string;
  lastUpdatedAt: string;
  createdBy: number;
  lastUpdatedBy: number;
  lastProgramUpdate?: string | null;
  code: string;
  status: number;
  lastUpdatedProgram: string;
  businessPlanId: number;
  exchangeRateBuy: number;
  exchangeRateSell: number;
  shippingMethod: string;
  packingMethod: string;
  weightPerContainer: number;
  estimatedTotalContainers: number;
  estimatedTotalBookings: number;
};

export type BusinessPlan = {
  id: string;
  code: string;
  draftPoId: number;
  contractId: string;
  contractCode?: string | null;
  codeBooking: number;
  totalQuantity: number;
  expectedPrice: number;
  currency: string;
  isApprove: boolean;
  totalRevenueExcludingVat: number;
  breakEvenPrice: number;
  actualBusinessProfit: number;
  profitMarginPercentage: number;
  customerName: string;
  goodName: string;
  createdAt: string;
  lastUpdatedAt?: string;
  createdBy?: number;
  lastUpdatedBy?: number;
  status: number;
  draftPo?: DraftPo | null;
  businessPlanCostItem?: BusinessPlanCostItem | null;
  businessPlanSupplierItems?: BusinessPlanSupplierItem[];
  businessPlanTransactionInfoItem?: BusinessPlanTransactionInfoItem | null;
};